Add more customization for websocket handshake
