+++ src/Octane.php
@@ -19,7 +19,7 @@ class Octane
      */
     public function table(string $table): Table
     {
-        $serverClass = config('octane.swoole.enableWebSocket', false)
+        $serverClass = config('octane.swoole.enable_web_socket', false)
             ? \Swoole\Websocket\Server::class
             : \Swoole\Http\Server::class;
