+++ src/OctaneServiceProvider.php
@@ -87,7 +87,7 @@ class OctaneServiceProvider extends ServiceProvider
         });
 
         $this->app->bind(DispatchesCoroutines::class, function ($app) {
-            $serverClass = config('octane.swoole.enableWebSocket', false)
+            $serverClass = config('octane.swoole.enable_web_socket', false)
                 ? \Swoole\Websocket\Server::class
                 : \Swoole\Http\Server::class;
