+++ src/Swoole/Handlers/OnWebSocketHandshake.php
@@ -2,36 +2,17 @@
 
 namespace Laravel\Octane\Swoole\Handlers;
 
-use Laravel\Octane\Swoole\WorkerState;
 use Swoole\Http\Request;
 use Swoole\Http\Response;
-use Swoole\WebSocket\Server;
 
 class OnWebSocketHandshake
 {
-    public function __construct(
-        protected Server $server,
-        protected array $serverState,
-        protected WorkerState $workerState
-    ) {
-    }
-
-    /**
-     * Handle the "handshake" Swoole event.
-     */
-    public function __invoke(Request $request, Response $response): void
-    {
-        if ($this->handshake($request, $response)) {
-            $this->workerState->worker->handleWebSocketOpen($this->server, $request);
-        }
-    }
-
     /**
      * Handle the handshake.
      *
      * @see https://www.swoole.co.uk/docs/modules/swoole-websocket-server
      */
-    protected function handshake(Request $request, Response $response): bool
+    public function handle(Request $request, Response $response): bool
     {
         $secWebSocketKey = $request->header['sec-websocket-key'];
